# perf: optimize JsonpJSON rendering to reduce allocations